re-structured configuration and baseUrl handling
