+++ public/index.html
@@ -2,8 +2,8 @@
 <html lang="en">
   <head>
     <!-- <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/draft-js/0.10.0/Draft.min.css"/> -->
-
     <meta charset="utf-8" />
+    <base href="__BASE_URL_TOKEN__" />
     <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta
       name="viewport"
@@ -27,9 +27,7 @@
     <title>HKube - Dashboard 🐡</title>
   </head>
   <body>
-    <noscript>
-      You need to enable JavaScript to run this app.
-    </noscript>
+    <noscript> You need to enable JavaScript to run this app. </noscript>
     <div id="root"></div>
     <!--
       This HTML file is a template.
